requirements.txt
Flask
flask==3.1.1
ultralytics==8.3.160
torch==2.2.2
opencv-python-headless==4.11.0.86
numpy==1.25.2
matplotlib==3.10.3
pillow==11.2.1
pyyaml==6.0.2
requests==2.32.4
scipy==1.15.3
torchvision==0.17.2
tqdm==4.67.1
psutil==7.0.0
py-cpuinfo==9.0.0
pandas==2.3.0
ultralytics-thop==2.0.14
flask-cors==4.0.0